Gryps/gryps_woven/server/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
Faker==15.0.0
faker-commerce==1.0.3
psycopg2==2.9.3
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
